NXP-22217: move BlobInfo to its own class
