+++ nuxeo-liveconnect-box-core/src/test/java/org/nuxeo/ecm/liveconnect/box/BoxTestCase.java
@@ -22,7 +22,7 @@ import java.util.UUID;
 
 import org.junit.Ignore;
 import org.junit.runner.RunWith;
-import org.nuxeo.ecm.core.blob.BlobManager.BlobInfo;
+import org.nuxeo.ecm.core.blob.BlobInfo;
 import org.nuxeo.ecm.core.blob.SimpleManagedBlob;
 import org.nuxeo.ecm.core.test.CoreFeature;
 import org.nuxeo.runtime.test.runner.Deploy;
