+++ nuxeo-liveconnect-core/src/main/java/org/nuxeo/ecm/liveconnect/core/AbstractLiveConnectBlobProvider.java
@@ -40,7 +40,7 @@ import org.nuxeo.ecm.core.api.DocumentModel;
 import org.nuxeo.ecm.core.api.NuxeoException;
 import org.nuxeo.ecm.core.api.repository.RepositoryManager;
 import org.nuxeo.ecm.core.blob.AbstractBlobProvider;
-import org.nuxeo.ecm.core.blob.BlobManager.BlobInfo;
+import org.nuxeo.ecm.core.blob.BlobInfo;
 import org.nuxeo.ecm.core.blob.BlobProvider;
 import org.nuxeo.ecm.core.blob.DocumentBlobProvider;
 import org.nuxeo.ecm.core.blob.ManagedBlob;
