+++ nuxeo-liveconnect-dropbox-core/src/test/java/org/nuxeo/ecm/liveconnect/dropbox/TestDropboxBlobProvider.java
@@ -30,8 +30,8 @@ import javax.inject.Inject;
 import org.apache.commons.io.IOUtils;
 import org.junit.Test;
 import org.nuxeo.ecm.core.api.Blob;
+import org.nuxeo.ecm.core.blob.BlobInfo;
 import org.nuxeo.ecm.core.blob.BlobManager;
-import org.nuxeo.ecm.core.blob.BlobManager.BlobInfo;
 import org.nuxeo.ecm.core.blob.BlobProvider;
 import org.nuxeo.ecm.core.blob.SimpleManagedBlob;
 import org.nuxeo.runtime.test.runner.RuntimeHarness;
