+++ nuxeo-liveconnect-google-drive-core/src/test/java/org/nuxeo/ecm/liveconnect/google/drive/TestGoogleDriveDocumentUpdate.java
@@ -34,8 +34,8 @@ import org.apache.commons.lang3.StringUtils;
 import org.junit.Test;
 import org.nuxeo.ecm.core.api.CoreSession;
 import org.nuxeo.ecm.core.api.DocumentModel;
+import org.nuxeo.ecm.core.blob.BlobInfo;
 import org.nuxeo.ecm.core.blob.BlobManager;
-import org.nuxeo.ecm.core.blob.BlobManager.BlobInfo;
 import org.nuxeo.ecm.core.blob.SimpleManagedBlob;
 import org.nuxeo.ecm.core.work.api.WorkManager;
 import org.nuxeo.ecm.liveconnect.update.BatchUpdateBlobProvider;
